+++ packages/website/components/account/storageManager/storageManager.js
@@ -8,15 +8,9 @@ import Button, { ButtonVariant } from 'components/button/button';
 import { useUser } from 'components/contexts/userContext';
 import { elementIsInViewport } from 'lib/utils';
 
-// Tiers available
-export const StorageTiers = {
-  TIER_1: '0',
-  TIER_2: '1',
-  TIER_3: '2',
-};
-
 // Raw TiB number of bytes, to be used in calculations
 const tebibyte = 1099511627776;
+const defaultStorageLimit = tebibyte;
 
 /**
  * @typedef {Object} StorageManagerProps
@@ -34,39 +28,23 @@ const mailTo = `mailto:${emailContent.mail}?subject=${emailContent.subject}&body
  * @returns
  */
 const StorageManager = ({ className = '', content }) => {
-  const storageTier = StorageTiers.TIER_1; // No tier available?
   const {
     storageData: { data, isLoading },
   } = useUser();
   const uploaded = useMemo(() => data?.usedStorage?.uploaded || 0, [data]);
   const psaPinned = useMemo(() => data?.usedStorage?.psaPinned || 0, [data]);
+  const limit = useMemo(() => data?.storageLimitBytes || defaultStorageLimit, [data]);
   const [componentInViewport, setComponentInViewport] = useState(false);
   const storageManagerRef = useRef(/** @type {HTMLDivElement | null} */ (null));
 
   const { maxSpaceLabel, unlockLabel, percentUploaded, percentPinned } = useMemo(
-    () =>
-      // Storage information by tier
-      ({
-        [StorageTiers.TIER_1]: {
-          maxSpaceLabel: content.tiers[0].max_space_label,
-          unlockLabel: content.tiers[0].unlock_label,
-          percentUploaded: (uploaded / tebibyte) * 100,
-          percentPinned: (psaPinned / tebibyte) * 100,
-        },
-        [StorageTiers.TIER_2]: {
-          maxSpaceLabel: content.tiers[1].max_space_label,
-          unlockLabel: content.tiers[1].unlock_label,
-          percentUploaded: (uploaded / (tebibyte * 10)) * 100,
-          percentPinned: (psaPinned / (tebibyte * 10)) * 100,
-        },
-        [StorageTiers.TIER_3]: {
-          maxSpaceLabel: `${Math.floor(uploaded + psaPinned / (tebibyte * 10) + 1) + content.tiers[2].max_space_label}`,
-          // every increment of 10 changes the amount of space used
-          percentUploaded: ((uploaded % (tebibyte * 10)) / (tebibyte * 10)) * 100,
-          percentPinned: ((psaPinned % (tebibyte * 10)) / (tebibyte * 10)) * 100,
-        },
-      }[storageTier]),
-    [storageTier, uploaded, psaPinned, content.tiers]
+    () => ({
+      maxSpaceLabel: `${Math.floor(limit / tebibyte)} ${content.max_space_tib_label}`,
+      unlockLabel: content.unlock_label,
+      percentUploaded: (uploaded / limit) * 100,
+      percentPinned: (psaPinned / limit) * 100,
+    }),
+    [uploaded, psaPinned, limit, content]
   );
 
   useEffect(() => {